Set up taxjar:transactions:sync CLI command
